consolidated shared code for LoginCpp
