+++ samples/LoginCpp/LoginCpp/LoginCpp.Windows/Dialogs.xaml.cpp
@@ -14,11 +14,6 @@
 //
 //******************************************************************************
 
-//
-// Dialogs.xaml.cpp
-// Implementation of the Dialogs class
-//
-
 #include "pch.h"
 #include "Dialogs.xaml.h"
 
@@ -37,80 +32,9 @@ using namespace Windows::UI::Xaml::Input;
 using namespace Windows::UI::Xaml::Media;
 using namespace Windows::UI::Xaml::Navigation;
 
-// The Blank Page item template is documented at http://go.microsoft.com/fwlink/?LinkId=234238
-
 Dialogs::Dialogs()
 {
     InitializeComponent();
-}
-
-
-void LoginCpp::Dialogs::FeedDialogButton_Click(Platform::Object^ sender, Windows::UI::Xaml::RoutedEventArgs^ e)
-{
-    FBSession^ s = FBSession::ActiveSession;
-
-    if (!s->LoggedIn)
-    {
-        OutputDebugString(L"The user is no longer logged in.\n");
-    }
-    else
-    {
-        PropertySet^ params = ref new PropertySet();
-        params->Insert(L"caption", L"I love Brussels Sprouts!");
-        params->Insert(L"link", L"https://en.wikipedia.org/wiki/Brussels_sprout");
-        params->Insert(L"description", L"Om Nom Nom!");
-
-        create_task(s->ShowFeedDialogAsync(params))
-            .then([=](FBResult^ Response)
-        {
-            OutputDebugString(L"Showed 'Feed' dialog.\n");
-        });
-    }
-}
-
-
-void LoginCpp::Dialogs::AppRequestsButton_Click(Platform::Object^ sender, Windows::UI::Xaml::RoutedEventArgs^ e)
-{
-    FBSession^ s = FBSession::ActiveSession;
-
-    if (!s->LoggedIn)
-    {
-        OutputDebugString(L"The user is no longer logged in.\n");
-    }
-    else
-    {
-        PropertySet^ params = ref new PropertySet();
-
-        params->Insert(L"title", L"I love Brussels Sprouts!");
-        params->Insert(L"message", L"Om Nom Nom!");
-
-        create_task(s->ShowRequestsDialogAsync(params))
-            .then([=](FBResult^ Response)
-        {
-            OutputDebugString(L"Showed 'Requests' dialog.\n");
-        });
-    }
-}
-
-void LoginCpp::Dialogs::SendDialogButton_Click(Platform::Object^ sender, Windows::UI::Xaml::RoutedEventArgs^ e)
-{
-    FBSession^ s = FBSession::ActiveSession;
-    if (!s->LoggedIn)
-    {
-        OutputDebugString(L"The user is no longer logged in.\n");
-    }
-    else
-    {
-        PropertySet^ params = ref new PropertySet();
-
-        params->Insert(L"link", L"http://example.com");
-
-        create_task(s->ShowSendDialogAsync(params))
-            .then([=](FBResult^ DialogResponse)
-        {
-            OutputDebugString(L"Showed 'Send' dialog.\n");
-        });
-    }
 }
 
 void LoginCpp::Dialogs::BackButton_Click(Platform::Object^ sender, Windows::UI::Xaml::RoutedEventArgs^ e)
