+++ samples/LoginCpp/LoginCpp/LoginCpp.Windows/UserInfo.xaml.cpp
@@ -74,51 +74,6 @@ Common::NavigationHelper^ UserInfo::NavigationHelper::get()
     return safe_cast<Common::NavigationHelper^>(GetValue(_navigationHelperProperty));
 }
 
-#pragma region Navigation support
-
-DependencyProperty^ UserInfo::_navigationHelperProperty =
-DependencyProperty::Register("NavigationHelper",
-TypeName(Common::NavigationHelper::typeid), TypeName(UserInfo::typeid), nullptr);
-
-/// The methods provided in this section are simply used to allow
-/// NavigationHelper to respond to the page's navigation methods.
-/// 
-/// Page specific logic should be placed in event handlers for the  
-/// <see cref="NavigationHelper::LoadState"/>
-/// and <see cref="NavigationHelper::SaveState"/>.
-/// The navigation parameter is available in the LoadState method 
-/// in addition to page state preserved during an earlier session.
-
-void UserInfo::OnNavigatedTo(NavigationEventArgs^ e)
-{
-	NavigationHelper->OnNavigatedTo(e);
-    FBSession^ sess = FBSession::ActiveSession;
-    if (sess->LoggedIn)
-    {
-        FBUser^ user = sess->User;
-        if (user)
-        {
-            UserId->Text = user->Id;
-            UserFirstName->Text = user->FirstName;
-            UserGender->Text = user->Gender;
-            UserLastName->Text = user->LastName;
-            UserLink->Text = user->Link;
-            UserLocale->Text = user->Locale;
-            UserName->Text = user->Name;
-            UserTimezone->Text = user->Timezone.ToString();
-            UserUpdatedTime->Text = user->UpdatedTime;
-            UserVerified->Text = user->Verified.ToString();
-            SquarePicture->UserId = user->Id;
-        }
-    }
-}
-
-void UserInfo::OnNavigatedFrom(NavigationEventArgs^ e)
-{
-	NavigationHelper->OnNavigatedFrom(e);
-}
-
-#pragma endregion
 
 /// <summary>
 /// Populates the page with content passed during navigation. Any saved state is also
@@ -149,15 +104,6 @@ void UserInfo::SaveState(Object^ sender, Common::SaveStateEventArgs^ e){
     (void)sender;	// Unused parameter
     (void)e; // Unused parameter
 }
-
-void LoginCpp::UserInfo::UserLikesButton_Click(Platform::Object^ sender, Windows::UI::Xaml::RoutedEventArgs^ e)
-{
-    LoginCpp::App^ a = dynamic_cast<LoginCpp::App^>(Application::Current);
-    Windows::UI::Xaml::Controls::Frame^ f = a->CreateRootFrame();
-    f->Navigate(UserLikes::typeid);
-}
-
-
 
 void LoginCpp::UserInfo::BackButton_Click(Platform::Object^ sender, Windows::UI::Xaml::RoutedEventArgs^ e)
 {
